Fixes optional character strings and characters(len=n)
Optional characters are just passed as a NULL char pointer,
without the byte at the end of the argument list.

We need to make sure to save the symbol_ref if self._value
depends on other variables so that we can re-resolve things
the next time we call a function.